move "Look & Feel/Edit_metadata" to its own widget
